Store direct pointer to delay_t in delay_control
No idea how it could possibly compile before, but delay_control needs a
pointer to the timer_mux, which it lacked, and which would make it a bit
too big and inefficient. Instead just store a pointer to the delay_t it
controls, which also makes more sense when you think about it.