Fix bootloop when -O > 0
Without -fno-toplevel-reorder, the custom asm in startup.hxx results in
uc_main not being placed right after the initialization code. The stock asm
handles this with a rcall into main, but that wastes a few bytes.

On the other hand, since rawr does not use globals (which disabling reorder
causes to not be optimized away even if possible, as a UB compatibility
measure), there doesn’t seem to be much of a drawback from using this flag.